+++ eslint.config.mjs
@@ -7,6 +7,7 @@ export default [
       '@typescript-eslint/no-explicit-any': 'off',
       '@typescript-eslint/no-param-reassign': 'off',
       'import/prefer-default-export': 'off',
+      'import/no-cycle': 'off',
     },
     settings: {
       'import/resolver': {
